next-app/requirements.txt
fastapi[all]>=0.135.1
google-genai>=1.66.0
dotenv>=0.9.9
